Codegen: reorganize symbols in `models` directory
