+++ godot-codegen/src/lib.rs
@@ -10,10 +10,8 @@ mod class_generator;
 mod codegen_special_cases;
 mod context;
 mod conv;
-mod domain_mapping;
-mod domain_models;
 mod interface_generator;
-mod json_models;
+mod models;
 mod special_cases;
 mod util;
 mod utilities_generator;
@@ -21,27 +19,22 @@ mod utilities_generator;
 #[cfg(test)]
 mod tests;
 
-use central_generator::{
-    generate_core_central_file, generate_core_mod_file, generate_sys_central_file,
-    generate_sys_classes_file,
-};
-use class_generator::{
-    generate_builtin_class_files, generate_class_files, generate_native_structures_files,
-};
-use context::Context;
-use interface_generator::generate_sys_interface_file;
-use json_models::{load_extension_api, JsonExtensionApi};
-use util::ident;
-use utilities_generator::generate_utilities_file;
-
 use crate::central_generator::{
-    generate_sys_builtin_lifecycle_file, generate_sys_builtin_methods_file,
+    generate_core_central_file, generate_core_mod_file, generate_sys_builtin_lifecycle_file,
+    generate_sys_builtin_methods_file, generate_sys_central_file, generate_sys_classes_file,
     generate_sys_utilities_file,
 };
-use crate::context::NotificationEnum;
-use crate::domain_models::ExtensionApi;
+use crate::class_generator::{
+    generate_builtin_class_files, generate_class_files, generate_native_structures_files,
+};
+use crate::context::{Context, NotificationEnum};
+use crate::interface_generator::generate_sys_interface_file;
+use crate::models::domain::{ExtensionApi, TyName};
+use crate::models::json::{load_extension_api, JsonExtensionApi};
+use crate::util::ident;
+use crate::utilities_generator::generate_utilities_file;
 use proc_macro2::{Ident, TokenStream};
-use quote::{quote, ToTokens};
+use quote::ToTokens;
 use std::path::{Path, PathBuf};
 
 pub type SubmitFn = dyn FnMut(PathBuf, TokenStream);
@@ -153,139 +146,6 @@ pub fn generate_core_files(core_gen_path: &Path) {
     watch.record("generate_native_structures_files");
 
     watch.write_stats_to(&core_gen_path.join("codegen-stats.txt"));
-}
-
-// ----------------------------------------------------------------------------------------------------------------------------------------------
-// Shared utility types
-
-// Same as above, without lifetimes
-
-#[derive(Clone, Eq, PartialEq, Hash, Debug)]
-struct GodotTy {
-    ty: String,
-    meta: Option<String>,
-}
-
-// impl GodotTy {
-//     fn new<'a>(ty: &'a String, meta: &'a Option<String>) -> Self {
-//         Self {
-//             ty: ty.clone(),
-//             meta: meta.clone(),
-//         }
-//     }
-// }
-
-// ----------------------------------------------------------------------------------------------------------------------------------------------
-
-#[derive(Clone, Debug)]
-pub enum RustTy {
-    /// `bool`, `Vector3i`
-    BuiltinIdent(Ident),
-
-    /// `Array<i32>`
-    BuiltinArray(TokenStream),
-
-    /// C-style raw pointer to a `RustTy`.
-    RawPointer { inner: Box<RustTy>, is_const: bool },
-
-    /// `Array<Gd<PhysicsBody3D>>`
-    EngineArray {
-        tokens: TokenStream,
-        #[allow(dead_code)] // only read in minimal config
-        elem_class: String,
-    },
-
-    /// `module::Enum`
-    EngineEnum {
-        tokens: TokenStream,
-        /// `None` for globals
-        #[allow(dead_code)] // only read in minimal config
-        surrounding_class: Option<String>,
-    },
-
-    /// `module::Bitfield`
-    EngineBitfield {
-        tokens: TokenStream,
-        /// `None` for globals
-        #[allow(dead_code)] // only read in minimal config
-        surrounding_class: Option<String>,
-    },
-
-    /// `Gd<Node>`
-    EngineClass {
-        /// Tokens with full `Gd<T>`
-        tokens: TokenStream,
-        /// only inner `T`
-        #[allow(dead_code)] // only read in minimal config
-        inner_class: Ident,
-    },
-}
-
-impl RustTy {
-    pub fn return_decl(&self) -> TokenStream {
-        match self {
-            Self::EngineClass { tokens, .. } => quote! { -> Option<#tokens> },
-            other => quote! { -> #other },
-        }
-    }
-}
-
-impl ToTokens for RustTy {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        match self {
-            RustTy::BuiltinIdent(ident) => ident.to_tokens(tokens),
-            RustTy::BuiltinArray(path) => path.to_tokens(tokens),
-            RustTy::RawPointer {
-                inner,
-                is_const: true,
-            } => quote! { *const #inner }.to_tokens(tokens),
-            RustTy::RawPointer {
-                inner,
-                is_const: false,
-            } => quote! { *mut #inner }.to_tokens(tokens),
-            RustTy::EngineArray { tokens: path, .. } => path.to_tokens(tokens),
-            RustTy::EngineEnum { tokens: path, .. } => path.to_tokens(tokens),
-            RustTy::EngineBitfield { tokens: path, .. } => path.to_tokens(tokens),
-            RustTy::EngineClass { tokens: path, .. } => path.to_tokens(tokens),
-        }
-    }
-}
-
-// ----------------------------------------------------------------------------------------------------------------------------------------------
-
-/// Contains multiple naming conventions for types (classes, builtin classes, enums).
-// TODO(bromeon, 2023-09): see if it makes sense to unify this with TypeNames (which is mostly used in central generator)
-#[derive(Clone, Eq, PartialEq, Hash)]
-pub struct TyName {
-    godot_ty: String,
-    rust_ty: Ident,
-}
-
-impl TyName {
-    fn from_godot(godot_ty: &str) -> Self {
-        Self {
-            godot_ty: godot_ty.to_owned(),
-            rust_ty: ident(&conv::to_pascal_case(godot_ty)),
-        }
-    }
-
-    fn description(&self) -> String {
-        if self.rust_ty == self.godot_ty {
-            self.godot_ty.clone()
-        } else {
-            format!("{}  [renamed {}]", self.godot_ty, self.rust_ty)
-        }
-    }
-
-    fn virtual_trait_name(&self) -> String {
-        format!("I{}", self.rust_ty)
-    }
-}
-
-impl ToTokens for TyName {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        self.rust_ty.to_tokens(tokens)
-    }
 }
 
 // ----------------------------------------------------------------------------------------------------------------------------------------------
